moving notebook snippet to notebook level
